requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
psycopg2-binary
sqlalchemy
python-multipart
python-magic
Pillow
cloudinary
yt-dlp
